training_backend/pyproject.toml
[project]
name = "training-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.35.0",
]
